Update the check for in use credentials
