+++ frontend/eda/access/credentials/hooks/useDeleteCredentials.tsx
@@ -4,7 +4,7 @@ import { compareStrings } from '../../../../../framework';
 import { useNameColumn } from '../../../../common/columns';
 import { requestDelete } from '../../../../common/crud/Data';
 import { idKeyFn } from '../../../../common/utils/nameKeyFn';
-import { InUseResources } from '../../../common/EdaResourcesComon';
+import { InUseResourceRefs } from '../../../common/EdaResourcesCommon';
 import { edaAPI } from '../../../common/eda-utils';
 import { EdaCredential } from '../../../interfaces/EdaCredential';
 import { useCredentialColumns } from './useCredentialColumns';
@@ -18,13 +18,12 @@ export function useDeleteCredentials(onComplete?: (credentials: EdaCredential[])
   const bulkAction = useEdaBulkConfirmation<EdaCredential>();
   return useCallback(
     async (credentials: EdaCredential[]) => {
-      const inUseDes = await InUseResources(credentials, edaAPI`/activations/?eda_credential_id=`);
+      const inUseDes = await InUseResourceRefs(credentials, edaAPI`/eda-credentials`);
       const inUseMessage =
         inUseDes && inUseDes.length > 0
           ? [t(`The following credentials are in use: ${inUseDes.join()}`)]
           : [];
       const forceParameter = inUseMessage.length > 0 ? '?force=true' : '';
-
       bulkAction({
         title: t('Permanently delete credentials', { count: credentials.length }),
         confirmText: t('Yes, I confirm that I want to delete these {{count}} credentials.', {
